backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via pytest
certifi==2020.12.5
    # via elasticsearch
coverage==5.5
    # via pytest-cov
django-cors-headers==3.7.0
    # via -r requirements.in
django-elasticsearch-dsl==7.1.4
    # via
    #   -r requirements.in
    #   django-elasticsearch-dsl-drf
django-filter==2.4.0
    # via -r requirements.in
django-nine==0.2.4
    # via django-elasticsearch-dsl-drf
django==3.1.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-nine
    #   djangorestframework
django_elasticsearch_dsl_drf==0.21
    # via -r requirements.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   django-elasticsearch-dsl-drf
elasticsearch-dsl==7.3.0
    # via
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
elasticsearch==7.11.0
    # via
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
iniconfig==1.1.1
    # via pytest
markdown==3.3.4
    # via -r requirements.in
mirakuru==2.3.0
    # via pytest-postgresql
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
port-for==0.4
    # via pytest-postgresql
psutil==5.8.0
    # via mirakuru
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements.in
pytest-django==4.1.0
    # via -r requirements.in
pytest-env==0.6.2
    # via -r requirements.in
pytest-postgresql==2.6.1
    # via -r requirements.in
pytest==6.2.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-postgresql
python-dateutil==2.8.1
    # via elasticsearch-dsl
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2021.1
    # via django
six==1.15.0
    # via
    #   django-elasticsearch-dsl
    #   django-elasticsearch-dsl-drf
    #   elasticsearch-dsl
    #   python-dateutil
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
urllib3==1.26.3
    # via elasticsearch
